api/requirements.txt
Flask==2.0.2
flask_cors==3.0.10
Werkzeug==2.0.2
pymongo==3.12.1
PyJWT==2.3.0
python-dotenv==0.19.2
marshmallow==3.14.1
webdriver-manager==3.5.2
beautifulsoup4==4.10.0
selenium==4.1.0
requests==2.26.0
openpyxl==3.0.9